--- a/.busted
+++ b/.busted
@@ -8,3 +8,5 @@ if ci then
     }
   }
 end
+
+return { }