--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
 /lua-resty-*-*
 !*.rockspec
+lua_modules