--- a/Makefile
+++ b/Makefile
@@ -1,2 +1,9 @@
+
+Roverfile.lock: Roverfile
+	rover lock
+
+lua_modules: Roverfile.lock Roverfile
+	rover install
+
 test:
 	@resty -e "require('busted.runner')({ standalone = false })"