--- a/Roverfile
+++ b/Roverfile
@@ -1,6 +1,7 @@
 luarocks {
     group 'test' {
         module { 'busted' },
+        module { 'benchmark-ips' },
     },
 }