# Before
```
Warming up --------------------------------------
           parse_uri
    277470 i/100ms

               split
    159546 i/100ms

Calculating -------------------------------------
  3494402.5 (±3.2%) i/s -   34961220 in  10.015314s
  1747467.2 (±4.1%) i/s -   17550060 in  10.060768s
```

# After 
```
Warming up --------------------------------------
           parse_uri
    269922 i/100ms

               split
    256759 i/100ms

Calculating -------------------------------------
  3126502.1 (±6.6%) i/s -   31310952 in  10.066162s
  2952137.4 (±3.7%) i/s -   29527285 in  10.016740s
```

# Conclustion

Almost 2x improvement.